fix: 주문 하기 API, 장바구니와 연동
fix: 주문 하기 API, 장바구니와 연동

+++ src/main/java/com/team/saver/market/order/dto/OrderDetailResponse.java
@@ -28,13 +28,13 @@ public class OrderDetailResponse {
 
     private String orderNumber;
 
-    private int orderPrice;
+    private long orderPrice;
 
     private int discountAmount;
 
     private PaymentType paymentType;
 
-    private int finalPrice;
+    private long finalPrice;
 
     private List<OrderMenuResponse> orderMenus;
